moved json reader to its own file
Ideally this would be a parser, but we don't have a great way of passing
arguments into parsers at the moment.

Might do that also after talking to Jerry.
