+++ gpt_index/__init__.py
@@ -69,6 +69,7 @@ from gpt_index.readers import (
     FaissReader,
     GithubRepositoryReader,
     GoogleDocsReader,
+    JSONReader,
     MboxReader,
     NotionPageReader,
     ObsidianReader,
@@ -76,7 +77,6 @@ from gpt_index.readers import (
     QdrantReader,
     RssReader,
     SimpleDirectoryReader,
-    SimpleJSONReader,
     SimpleMongoReader,
     SimpleWebPageReader,
     SlackReader,
@@ -126,7 +126,7 @@ __all__ = [
     "ObsidianReader",
     "Document",
     "SimpleDirectoryReader",
-    "SimpleJSONReader",
+    "JSONReader",
     "SimpleMongoReader",
     "NotionPageReader",
     "GoogleDocsReader",
