+++ gpt_index/readers/__init__.py
@@ -16,11 +16,12 @@ from gpt_index.readers.elasticsearch import ElasticsearchReader
 from gpt_index.readers.faiss import FaissReader
 
 # readers
-from gpt_index.readers.file.base import SimpleDirectoryReader, SimpleJSONReader
+from gpt_index.readers.file.base import SimpleDirectoryReader
 from gpt_index.readers.github_readers.github_repository_reader import (
     GithubRepositoryReader,
 )
 from gpt_index.readers.google_readers.gdocs import GoogleDocsReader
+from gpt_index.readers.json import JSONReader
 from gpt_index.readers.make_com.wrapper import MakeWrapper
 from gpt_index.readers.mbox import MboxReader
 from gpt_index.readers.mongo import SimpleMongoReader
@@ -46,7 +47,7 @@ __all__ = [
     "WikipediaReader",
     "YoutubeTranscriptReader",
     "SimpleDirectoryReader",
-    "SimpleJSONReader",
+    "JSONReader",
     "SimpleMongoReader",
     "NotionPageReader",
     "GoogleDocsReader",
